fixes final commit background tasks
