+++ training/background_tasks/tasks.py
@@ -1,5 +1,6 @@
 import time
 from rq import get_current_job
+from training.app import app, mail
 
 
 def example(seconds):
@@ -13,3 +14,13 @@ def example(seconds):
     job.meta['progress'] = 100
     job.save_meta()
     print('Task completed')
+
+
+def send_email_function(msg):
+    with app.app_context():
+        mail.send(msg)
+
+
+def send_email_async(msg):
+    job = app.task_queue.enqueue(send_email_function, msg)
+    job.get_id()
